Rework html_anchor for tabs
An optional tab id is provided to the GuitarTab
constructor.
If missing, a new id is generated from the url
(using hashlib instead of hash() to have the
same result on each run).
Then, the id is used along with the song name
to generate the html_anchor. This is now
performed from a single place.