Fix recent commit about ordering
30a15098617c487a4eeebff2bf72619a0eea90cf introduced additional
data in the by_xxx methods to ensure a consistent ordering.
However, these additional fields ended up being in the data
displayed in the index which makes no sense (which was hidden
in the middle of the other changes).

Removing the corresponding fix and doing things differently
brings back the wanted behavior.