Rewrite logic around Chords: specific content and comparison
Instead of setting fields in Chords classes and then
implement a get_specific_content() method relying on
these fields, we could compute the content directly
and pass it to the constructor.

This makes various things slightly better:
 - instead of struggling to compare object which
may or may not lead to similar HTML code, we can
directly compare the HTML code and forget about
most other details -> this leads to fewer duplicated
chords in the final HTML content
 - because we don't set additional fields after the
call to super().__init__, we don't need to explicly
call the register() method, it can be called directly
from the parent constructor.
 - because, we have the same common fields in class
subclasses, we don't need to override:
    * __eq__
	* __str__
 - eventually, we can also get rid of all specific
constructors
 - we can get rid of the pruning logic from previous
commit