Use a single Chords class
Recent changes suggested that maybe we did
not really need one subclass for each type
of Chords. Indeed, the different subclasses
of AbstractChords are just wrappers to
call the AbstractChords constructor with
the relevant arguments. Once, the instance
is built, there is no method overriden
whatsoever.

Thus, we could probably use a single
Chords class which does not really need
to be considered abstract anymore.

Also, this solves an issue with duplicated
chords. Indeed, recently, we made sure
that Chords objects were compared based
on how they'd end up being formatted
instead of the underlying data used to
get that formatting. This had the nice
benefit of removing similar duplicated chords
based on slightly different data. However,
this was not perfect yet: we could have
Chords which same formatting but that
would be considered different because they
are instance of different subclasses of
AbstractChords (and thus, the objects
should not be considered equal). Having
a single Chords class removes this issue
and gets rid of various duplicated chords
in the final HTML document.

It would probably make sense to do some
cleanup in the chords module and its
usage but it is probably better to split
this in a different commit.